Quote minus signs and dashes.
Unquoted - in roff means a hyphen, so when - is used in command-line
options, examples and URLs it needs to be quoted. The difference isn't
very obvious in a text man viewer, but if you format a man page as PDF
or HTML, then unquoted - looks different and won't copy as the right
character.