Fix ctype(3) issues.
The argument of these functions is of type int, but only a very
restricted subset of values are actually valid.  The argument
must either be the value of the macro EOF (which has a negative
value), or must be a non-negative value within the range
representable as unsigned char.  Passing invalid values leads
to undefined behavior.

NetBSD actively warns about these issues at compile-time
("warning: array subscript has type 'char'").  Casting to
(int) is incorrect, as it can incorrectly suppress the
warning for invalid values.  The arguments must be cast to
(unsigned char).

Remove inclusion of <ctype.h> from files that do not use it.